This is a slight rework of a post of Jeff Tchang at Return Boolean True
