--- FireEyeFi/chrome.manifest
@@ -1,2 +1,0 @@
-content     sample    chrome/content/
-overlay chrome://browser/content/browser.xul chrome://sample/content/sample.xul
