--- FireEyeFi/chrome/content/sample.xul
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-<overlay id="sample" 
-         xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
- <statusbar id="status-bar">
-  <statusbarpanel id="my-panel" label="Hello, World"  />
- </statusbar>
-</overlay>
