--- Release 2.0/DefaultSettings.ini
@@ -1,75 +1,0 @@
-# Main configuration
-
-[Global]
-
-# The directives in this section affect the overall operation
-# of the Eye-Fi server
-
-
-#
-# ListenPort: Allows you to bind the Eye-Fi server to a specific port.
-# 
-#ListenPort=59278
-
-#
-# ConsoleOutput: Logging can automatically be sent to the console.
-# Set this to False if you don't want any console output. Console
-# output is also considered stdout.
-#
-#ConsoleOutput=True
-
-
-#
-# LogFile: Controls where to write the Eye-Fi logs.
-#
-#LogFile=
-
-
-#
-# LogLevel: The level of verbosity in both the logs and the console
-# output. From most verbose to lease verbose the settings are
-# DEBUG, INFO, WARNING, ERROR or CRITICAL
-#
-LogLevel=DEBUG
-
-#
-# DownloadLocation: The directory in which to put the incoming pictures. By
-# default the pictures are put in a sub directory called "pictures" from where
-# the script is originally started.
-#
-#On Windows:
-#  This would set the downloads to a directory called pictures.
-#DownloadLocation=.\\pictures
-#
-#On Unix:
-#DownloadLocation=/tmp
-#
-DownloadLocation=.\\pictures
-
-
-
-#
-# ExecuteOnUpload: This parameter is used to define an external program or script to
-# execute after a file is uploaded. By nature this command is very dangerous and should
-# be used carefully. Enabling this command can serve as a means of compromising a
-# system or disclosing information via bugs in external programs or scripts.
-#
-# This parameter executes the specified command on each incoming file passing in the full
-# file path as the first argument. There is no default for this command (nothing is executed
-# if this parameter is left blank).
-# 
-#On Windows:
-#
-#ExecuteOnUpload=C:\\Windows\\system32\\mspaint.exe
-
-
-[Card]
-
-# The directives in this section affect the physical card settings
-
-#
-# UploadKey: The Eye-Fi upload key. It is in C:\Documents and Settings\
-# <User>\Application Data\Eye-Fi\Settings.xml. This needs to be set for
-# the server to function correctly.
-#
-UploadKey=c686e547e3728c63a8f78729c1592757
