+++ etc/eyefiserver.conf
@@ -1,0 +1,27 @@
+
+[EyeFiServer]
+
+# host name and port to listen on
+# you can leave hostname empty for localhost
+
+host_name:
+host_port:59278
+
+# To use this script you need to have your Eye-Fi upload key.
+# You can find it after configuring the card, 
+# which you can currently on do on windows or mac
+# It is inside C:\Documents and Settings\<User>\Application Data\Eye-Fi\Settings.xml on windows
+# or ~/Applications Data/Eye-Fi/Settings.xml on mac
+# search for it and paste it here:
+
+upload_key:774e46f14990456a9ab7b4d22f1da138
+
+# When connecting, all files are downloaded in one directory
+# the name of the directory can be a strftime formatted string like 
+# /home/myblog/pictures/%%Y-%%m-%%d
+# notice the double percent sign to escape % from ini interpolation
+
+upload_dir:/home/pike/Desktop/eyefifolder/%%Y/%%Y%%m%%d
+
+
+
